requirements.txt
flask-socketio==5.3.6
eventlet==0.33.3
flask==1.1.4
python-dotenv==1.0.0
pytz==2022.6
pandas==1.5.0
mysqlclient==2.2.0
html2image==2.0.4.3
Pillow==9.4.0
Jinja2==2.11.3
flask_sqlalchemy==2.5.1
Flask-Login==0.5.0
sqlalchemy==1.3.23
Werkzeug==0.16.1
MarkupSafe==1.1.1
xhtml2pdf==0.2.13
pdfrw==0.4
reportlab==4.0.7
numpy==1.22.2
Flask-Toastr==0.5.8